add device orientation provider
